now needs extensive testing, other endpoints added
