+++ pyzendesk/zenContainer.py
@@ -1,25 +1,61 @@
 from sortedcontainers import SortedList, SortedSet, SortedDict
 from Session import Session
 
-class zenList:
-    def __init__(self, data=None):
-        self.data = SortedDict(data)
+class zenCredentials():
 
-    def __repr__(self):
-        return repr(self.data)
+    base_url = 'https://{}.zendesk.com/api/v2'
 
+    def __init__(self, subdomain, user, password='', token=''):
+        if token:
+            password = token
+            user = user + '/token'
+        self.auth = (user, password)
+        self.subdomain = subdomain
+
+    def create_url(self, endpoint, params = None, page = None):
+        param_string = params or []
+        param_string = param_string if type(param_string) is list else [param_string]
+        if(page):
+            param_string.append('page={}'.format(page))
+        param_string = '&'.join(param_string)
+        url = '{}/{}?{}'.format(self.base_url, endpoint, param_string)
+        return url.format(self.subdomain)
+
+
+class Endpoint:
+    def __init__(self, member, wrapper):
+        self.member = member
+        self.wrapper = wrapper
+
+    def get_all(self, fetch=False):
+        data = self.wrapper.get_all(self.member, fetch)
+        return data
+
+    def get(self, object_id):
+        return self.wrapper.get(member, object_id)
 
     def __call__(self, key):
+        return self.get(key)
+
+    def endpoint(self, key):
+        return '{}/{}'.format(self.wrapper.all_endpoints[self.member]['plural'], key)
+
+    def __getitem__(self, key):
+        try:
+            self.count
+        except NameError:
+            self.count = session.count(self.endpoint(key))
         if type(key) == slice:
             start = key.start or 0
             stop = key.stop or 0
             step = key.step or 1
-            return [self.data[index] for index in self.data.keys() if index in range(start, stop, step)]
-        return self.data[key]
+            if(len(range(start, stop, step)) > self.count/100):
+                self.get_all(fetch=True)
+            return [self.get(index) for index in range(start, stop, step)]
+        return self.get(key)
 
-    def __getitem__(self, key):
-        return self.data.values()[key]
-
+    def __iter__(self):
+        return self.get_all()
 
 
 class zenContainer:
@@ -42,7 +78,7 @@ class zenContainer:
         self.url = self.wrapper.session.create_url(self.endpoint)
         self.get()
 
-    def update(self, data):
+    def update(self):
         if self.id:
             print('put')
             self.wrapper.session.put(self.endpoint, self.raw)
@@ -53,6 +89,12 @@ class zenContainer:
             self.wrapper.expire_cache(self.endpoint)
             self.wrapper.session.post(self.endpoint, self.raw)
             self.should_refresh = True
+
+    def delete(self):
+        if not self.id:
+            return
+        else:
+            self.wrapper.session.delete(self.id)
 
     def get(self, cache=True):
         if(self.should_refresh):
